chore: split main form for improved readability
