+++ libs/application/templates/health-insurance-declaration/src/forms/healthInsuranceDeclarationForm/index.ts
@@ -1,0 +1,33 @@
+import { buildForm } from '@island.is/application/core'
+import { Form, FormModes } from '@island.is/application/types'
+import Logo from '../../assets/Logo'
+import { applicantInfoSection } from './applicantInfoSection'
+import { noHealthInsuranceSection } from './noHealthInsuranceSection'
+import { studentOrTouristSection } from './studentOrTouristSection'
+import { registerPersonsSection } from './registerPersonsSection'
+import { touristResidencySection } from './touristResidencySection'
+import { studentResidencySection } from './studentResidencySection'
+import { educationalConfirmationSection } from './educationalConfirmationSection'
+import { dateSection } from './dateSection'
+import { overviewSection } from './overviewSection'
+import * as m from '../../lib/messages'
+
+export const HealthInsuranceDeclarationForm: Form = buildForm({
+  id: 'HealthInsuranceDeclarationDraft',
+  title: m.application.general.name,
+  renderLastScreenButton: true,
+  renderLastScreenBackButton: true,
+  logo: Logo,
+  mode: FormModes.DRAFT,
+  children: [
+    applicantInfoSection,
+    noHealthInsuranceSection,
+    studentOrTouristSection,
+    registerPersonsSection,
+    touristResidencySection,
+    studentResidencySection,
+    educationalConfirmationSection,
+    dateSection,
+    overviewSection,
+  ],
+})
